refactor: 메뉴 검증로직을 클래스 분리

+++ src/main/java/kitchenpos/domain/Money.java
@@ -40,7 +40,7 @@ public class Money {
         return value;
     }
 
-    public Money times(final int quantity) {
+    public Money times(final long quantity) {
         return new Money(value.multiply(BigDecimal.valueOf(quantity)));
     }
